refactor: change Car package and modify goForward
